Create Age and Fat
Suppose that a hospital tested the age and body fat data for 18 randomly selected adults with the following results: 
 
(a) Calculate the mean, median, and standard deviation of age and %fat. 
(b) Draw the boxplots for age and %fat.
(c) Draw a scatter plot and a q-q plot based on these two variables.